coroapi.h, confdb.h, objdb.h: big s/int/size_t/ change
* exec/mainconfig.c (objdb_get_string, objdb_get_int):
* exec/totemconfig.c (objdb_get_string, objdb_get_int)
(totem_config_keyread, totem_key_change_notify):
* include/corosync/confdb.h (confdb_callbacks_t):
* include/corosync/engine/coroapi.h (group_len, object_len)
(key_len, validate_callback, object_key_change_notify_fn_t)
(object_create_notify_fn_t, object_destroy_notify_fn_t)
(object_notify_callback_fn_t, object_reload_notify_fn_t)
(object_create, object_key_create, object_find_create)
(object_key_get, object_key_replace, object_key_delete)
(object_iter, object_key_iter, object_name_get)
(object_key_iter_from, object_key_increment)
(object_key_decrement):
* include/corosync/engine/objdb.h (object_key_change_notify_fn_t)
(object_create_notify_fn_t, object_destroy_notify_fn_t)
(object_len, key_len, validate_callback, object_create)
(object_key_create, object_find_create, object_key_get)
(object_key_replace, object_key_delete, object_iter)
(object_key_iter, object_name_get, object_key_iter_from)
(object_key_increment, object_key_decrement):
* lib/confdb.c (confdb_object_create, confdb_key_create)
(confdb_key_delete, confdb_key_get, confdb_key_increment)
(confdb_key_decrement, confdb_key_replace, confdb_object_find)
(confdb_object_iter, confdb_key_iter):
* lib/sa-confdb.c (confdb_sa_object_create, confdb_sa_key_create)
(confdb_sa_key_delete, confdb_sa_key_get)
(confdb_sa_key_increment, confdb_sa_key_decrement)
(confdb_sa_key_replace, confdb_sa_object_find)
(confdb_sa_object_iter, confdb_sa_key_iter):
* lib/sa-confdb.h:
* services/confdb.c (message_handler_req_lib_confdb_key_replace):
* services/votequorum.c (objdb_get_string, objdb_get_int)
(quorum_key_change_notify, votequorum_objdb_reload_notify):

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2040 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/corosync/confdb.h
@@ -154,7 +154,7 @@ cs_error_t confdb_object_create (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *object_name,
-	int object_name_len,
+	size_t object_name_len,
 	hdb_handle_t *object_handle);
 
 cs_error_t confdb_object_destroy (
@@ -173,17 +173,17 @@ cs_error_t confdb_key_create (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	const void *value,
-	int value_len);
+	size_t value_len);
 
 cs_error_t confdb_key_delete (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	const void *value,
-	int value_len);
+	size_t value_len);
 
 /*
  * Key queries
@@ -192,32 +192,32 @@ cs_error_t confdb_key_get (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	void *value,
-	int *value_len);
+	size_t *value_len);
 
 cs_error_t confdb_key_replace (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	const void *old_value,
-	int old_value_len,
+	size_t old_value_len,
 	const void *new_value,
-	int new_value_len);
+	size_t new_value_len);
 
 cs_error_t confdb_key_increment (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	unsigned int *value);
 
 cs_error_t confdb_key_decrement (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	unsigned int *value);
 
 /*
@@ -234,7 +234,7 @@ cs_error_t confdb_object_find (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	const void *object_name,
-	int object_name_len,
+	size_t object_name_len,
 	hdb_handle_t *object_handle);
 
 cs_error_t confdb_object_find_destroy(
@@ -250,7 +250,7 @@ cs_error_t confdb_object_iter (
 	hdb_handle_t parent_object_handle,
 	hdb_handle_t *object_handle,
 	void *object_name,
-	int *object_name_len);
+	size_t *object_name_len);
 
 cs_error_t confdb_object_iter_destroy(
 	confdb_handle_t handle,
@@ -267,9 +267,9 @@ cs_error_t confdb_key_iter (
 	confdb_handle_t handle,
 	hdb_handle_t parent_object_handle,
 	void *key_name,
-	int *key_name_len,
+	size_t *key_name_len,
 	void *value,
-	int *value_len);
+	size_t *value_len);
 
 /*
  * Get/set context variable
